feat: show 3 latest features with a link to the Features section in worklog empty state
Replace the single random-feature widget with the 3 most recent features
(reusing FeatureCard) plus a "Все возможности" link to /features. Drop the now
unused GetRandomFeatureAsync.
