--- src/Pet.Jira.Web/Components/Features/RandomFeatureViewer.razor
@@ -1,31 +1,0 @@
-@using Pet.Jira.Web.Components.Common
-@using Pet.Jira.Web.Components.Markdown
-
-@if (_isLoading)
-{
-    <!-- Reading the feature is fast; render nothing to avoid a flash of fallback content. -->
-}
-else if (_feature is null)
-{
-    <Stub Message="@FallbackMessage" />
-}
-else
-{
-    <div class="feat-widget" @onclick="OpenDetailAsync" style="cursor:pointer">
-        <div class="feat-widget__icon">
-            <MudIcon Icon="@Icons.Material.Outlined.AutoAwesome" />
-        </div>
-        <div>
-            <div class="feat-widget__eyebrow">А вы знали?</div>
-            <MudText Typo="Typo.h6" Class="mt-1" Style="font-weight:600">@_feature.Metadata.Title</MudText>
-            <div class="feat-widget__preview mt-1">
-                <MudMarkdown Value="@_feature.PreviewMarkdown" />
-            </div>
-            <MudButton Class="mt-2" Variant="Variant.Text" Color="Color.Primary" Size="Size.Small"
-                       EndIcon="@Icons.Material.Filled.ArrowForward"
-                       OnClick="OpenDetailAsync">
-                Подробнее
-            </MudButton>
-        </div>
-    </div>
-}
