+++ src/Pet.Jira.Web/Components/Worklogs/WorklogList.razor
@@ -60,7 +60,7 @@
 @if (Items.IsEmpty())
 {
     <MudPaper Class="pa-4 pet-paper" Elevation="0">
-        <RandomFeatureViewer FallbackMessage="To display a list of actual and estimated worklogs, set the search options and click the 'Search' button" />
+        <LatestFeaturesViewer FallbackMessage="To display a list of actual and estimated worklogs, set the search options and click the 'Search' button" />
     </MudPaper>
 }
 else
